refactor: 프로필 정보 표시 컴포넌트 구조 및 스타일 분리 (#49)
* refactor: ProfileInfo 구조 및 이미지 처리 로직 개선

* refactor: ProfileInfo 스타일 컴포넌트 단위로 분리
